Move format configs into own file (faster loads)
